Add files via upload
Create a Python program that will take a photo every time a button is pressed. The camera preview should be enabled for at least 2 seconds before the photo is taken.

The photos should never be overwritten. You will need to come up with a naming convention that will ensure that the name of each photo will be unique.

Add a timestamp to the actual photo that includes the precise date/time that the photo was taken.

Hint: you will want to look at the datetime Python framework

To submit, upload your source code to github and post a link to your repo.